#6349: Stazeni a rozbaleni souboru zip
Vytvoreni tridy Parser, ktera zatim pouze stahuje soubory a extrahuje je. Vytvorena slozka pro ulozeni komprimovanych souboru.
