--- app_old/app/view/footer.php
@@ -1,9 +1,0 @@
-                </div>
-            </div>
-            <div>
-                &copy; ASWI team 2018   
-            </div>
-        </div>
-    </body>
-  
-</html>
