--- app_old/app/view/header.php
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang='cs'>
-
-    <head>
-        <title>Průjezd vozidel - Plzeňský kraj</title>
-        <meta charset="utf-8" />
-        <meta name="description" content="" />
-        <meta name="keywords" content="" />
-        <meta name="robots" content="all" />
-        <link rel="stylesheet" href="../../css/main.css" type="text/css" />
-    </head>
-    
-    <body>
-        <div>
-            <h1>Zobrazení dat o průjezdu vozidel pro Plzeňský kraj</h1>
-
