--- app_old/app/view/menu.php
@@ -1,6 +1,0 @@
-﻿<div>
-    <div>
-        <a href="control.php">Úvodní strana</a>
-        <a href="control.php?view=xxx">Další odkaz</a>
-    </div>
-    <div>
