--- app_old/download/20180316/Locations.csv
@@ -1,103 +1,0 @@
-"Česká Kubice, směr od Německa"|"Česká Kubice"|"Česká Kubice"|"KP055"|"5"
-"Česká Kubice, směr od Babylonu"|"Česká Kubice"|"Česká Kubice"|"KP056"|"5"
-"Holýšov, směr od Plzně"|"Holýšov"|"Jiráskova třída"|"KP051"|"7"
-"Horšovský Týn, směr od Boru"|"Horšovský Týn"|"Masarykova"|"KP066"|"8"
-"Horšovský Týn, směr od Stříbra"|"Horšovský Týn"|"Gorkého"|"KP067"|"8"
-"Horšovský Týn, směr od Bělé"|"Horšovský Týn"|"Nová Ves"|"KP068"|"8"
-"Klenčí pod Čerchovem, směr od Německa"|"Klenčí pod Čerchovem"|"Klenčí pod Čerchovem"|"KP061"|"10"
-"Klenčí pod Čerchovem, směr od Domažlic"|"Klenčí pod Čerchovem"|"Klenčí pod Čerchovem"|"KP062"|"10"
-"Libkov, směr od Klatov"|"Libkov"|"Libkov"|"KP069"|"11"
-"Poběžovice, směr od Klenčí"|"Poběžovice"|"Slovanská"|"KP071"|"14"
-"Staňkov, směr od H.Týna"|"Staňkov"|"Americká"|"KP053"|"16"
-"Běšiny, směr od Klatov"|"Běšiny"|"Běšiny"|"KP072"|"18"
-"Běšiny, směr od l/27"|"Běšiny"|"Běšiny"|"KP073"|"18"
-"Červené Poříčí, směr od Plzně"|"Červené Poříčí"|"Červené Poříčí"|"KP059"|"19"
-"Dlouhá Ves, směr od Sušice"|"Dlouhá Ves"|"Dlouhá Ves"|"KP058"|"20"
-"Kolinec, směr od Sušice"|"Kolinec"|"Kolinec"|"KP057"|"25"
-"Benešovice, směr od Stříbra"|"Benešovice"|"Benenšovice"|"KP060"|"32"
-"Broumov, směr od Plané"|"Broumov"|"Broumov"|"KP065"|"34"
-"Broumov, směr od Německa"|"Broumov"|"Broumov"|"KP064"|"34"
-"Tachov, směr od Plané"|"Tachov"|"Sokolovská"|"KP063"|"42"
-"Tachov, směr od Plzně"|"Tachov"|"Plzeňská"|"KP052"|"42"
-"Tachov, směr od Halže"|"Tachov"|"Moravská"|"KP054"|"42"
-"Domažlice, směr od Klatov"|"Domažlice"|"Masarykova"|"KP081"|"6"
-"Hostouň, směr od Bělé"|"Hostouň"|"Petra Bezruče"|"KP091"|"9"
-"Mířkov, směr od Boru"|"Mířkov"|"Mířkov"|"KP070"|"12"
-"Nemanice, směr od Poběžovic"|"Nemanice"|"Nemanice"|"KP090"|"13"
-"Všeruby - Brůdek, směr od Kdyně"|"Všeruby"|"Brůdek"|"KP077"|"17"
-"Všeruby, směr od Německa"|"Všeruby"|"Všeruby"|"KP078"|"17"
-"Janovice nad Úhlavou, směr od Klatov"|"Janovice nad Úhlavou"|"Nýrská"|"KP075"|"22"
-"Janovice nad Úhlavou, směr od Německa"|"Janovice nad Úhlavou"|"Nýrská"|"KP076"|"22"
-"Mochtín - Kocourov, směr od Horažďovic"|"Mochtín"|"Kocourov"|"KP079"|"27"
-"Mochtín, směr od Klatov"|"Mochtín"|"Mochtín"|"KP080"|"27"
-"Plánice, směr od Klatov"|"Plánice"|"Klatovská"|"KP087"|"29"
-"Stráž - Bonětice, směr od H. Týna"|"Stráž - Bonětice"|"Stráž - Bonětice"|"KP074"|"40"
-"Bělá nad Radbuzou, směr od Německa"|"Bělá nad Radbuzou"|"Dlouhá"|"KP119"|"1"
-"Bělá nad Radbuzou, směr od H. Týna"|"Bělá nad Radbuzou"|"Nádražní"|"KP120"|"1"
-"Horažďovice, směr od Klatov"|"Horažďovice"|"Komenského"|"KP112"|"21"
-"Horažďovice, směr od Strakonic"|"Horažďovice"|"Strakonická"|"KP113"|"21"
-"Horažďovice, směr od Velkého Boru"|"Horažďovice"|"Plzeňská"|"KP114"|"21"
-"Kašperské Hory, směr od Vimperka "|"Kašperské Hory"|"Vimperská"|"KP102"|"23"
-"Modrava, směr od Srní"|"Modrava"|"Modrava"|"KP088"|"26"
-"Modrava - Filipova Huť, směr od Kvildy"|"Modrava"|"Modrava (Filipova Huť)"|"KP089"|"26"
-"Nýrsko Klostermannova"|"Nýrsko"|"Klostermanova"|"KP083"|"28"
-"Nýrsko, směr od Svaté Kateřiny"|"Nýrsko"|"Palackého"|"KP098"|"28"
-"Nýrsko, směr od Chudenína"|"Nýrsko"|"Komenského"|"KP099"|"28"
-"Nýrsko, směr od Ž.Rudy"|"Nýrsko"|"U Banky"|"KP084"|"28"
-"Sušice, směr od Dlouhé Vsi"|"Sušice"|"Dlouhovesská"|"KP131"|"31"
-"Sušice, směr od Petrovic"|"Sušice"|"Volšovská"|"KP132"|"31"
-"Sušice, směr od Horažďovic"|"Sušice"|"Nádražní"|"KP103"|"31"
-"Železná Ruda - Alžbětín, směr od Německa"|"Železná Ruda"|"Železná Ruda"|"KP092"|"2"
-"Železná Ruda - Špičák, směr od Nýrska"|"Železná Ruda"|"Železná Ruda"|"KP093"|"2"
-"Bor, směr od D5"|"Bor"|"Nádražní"|"KP094"|"33"
-"Chodová Planá, směr od Chebu"|"Chodová Planá"|"Pohraniční stráže"|"KP100"|"36"
-"Planá, směr do Plzně"|"Planá"|"Tachovská"|"KP101"|"37"
-"Planá, směr do Chebu"|"Planá"|"Tachovská"|"KP085"|"37"
-"Rozvadov, směr od Německa"|"Rozvadov"|"Rozvadov"|"KP095"|"39"
-"Rozvadov, směr od Přimdy"|"Rozvadov"|"Rozvadov"|"KP096"|"39"
-"Stříbro, směr od Boru"|"Stříbro"|"5. května"|"KP097"|"41"
-"Stříbro, směr od centra"|"Stříbro"|"Plzeňská"|"KP086"|"41"
-"Babylon ve směru od Německa"|"Babylon"|"Babylon"|"KP117"|"3"
-"Babylon ve směru od H.Týna"|"Babylon"|"Babylon"|"KP118"|"3"
-"Brnířov od Kdyně"|"Brnířov"|"Brnířov"|"KP082"|"4"
-"Bělá nad Radbuzou, směr od náměstí"|"Bělá nad Radbuzou"|"Náměstí"|"KP108"|"1"
-"Domažlice, směr od od H.Týna"|"Domažlice"|"P. Velikého"|"KP110"|"6"
-"Domažlice, směr od Draženova"|"Domažlice"|"28. října"|"KP111"|"6"
-"Horšovský Týn, směr na přechodu od Německa"|"Horšovský Týn"|"Na Přeložce"|"KP105"|"8"
-"Horšovský Týn, směr na přechodu od Plzně"|"Horšovský Týn"|"Nádražní"|"KP106"|"8"
-"Nemanice, směr od Německa Lísková"|"Nemanice"|"Nemanice - Lísková"|"KP121"|"13"
-"Pocinovice ve směru od Nýrska"|"Pocinovice"|"Pocinovice"|"KP122"|"15"
-"Klatovy - Štěpánovice, směr od Plzně"|"Klatovy"|"Štěpánovice"|"KP123"|"24"
-"Klatovy směr od Horažďovic"|"Klatovy"|"Puškinova"|"KP124"|"24"
-"Klatovy směr od Železné Rudy"|"Klatovy"|"5. května Luby"|"KP125"|"24"
-"Klatovy - Plzeňská, směr z centra"|"Klatovy"|"Plzeňská"|"KP126"|"24"
-"Klatovy - Domažlická, směr z centra"|"Klatovy"|"Domažlická"|"KP127"|"24"
-"Klatovy směr od Domažlic"|"Klatovy"|"Domažlická"|"KP128"|"24"
-"Klatovy směr od Dolan"|"Klatovy"|"Koldinova"|"KP129"|"24"
-"Klatovy směr od Plánice"|"Klatovy"|"Plánická"|"KP130"|"24"
-"Bor, směr od Přimdy"|"Bor"|"Přimdská"|"KP116"|"33"
-"Bor, II/200 Plzeňská"|"Bor"|"Plzeňská"|"KP107"|"33"
-"Přimda směr od Boru"|"Přimda"|"Borská"|"KP135"|"38"
-"Přimda směr od Rozvadova"|"Přimda"|"Borská"|"KP136"|"38"
-"Srní, směr od Rejštejna "|"Srní"|"Srní"|"KP115"|"30"
-"Halže - Branka, II/199 od Německa"|"Halže - Branka"|"Halže - Branka"|"KP104"|"35"
-"Horní Folmava, I/26, směr od Německa"|"Horní Folmava"|"Horní Folmava"|"KP109"|"5"
-"Sušice, Hrádecká ul. u nákupních center"|"Sušice"|"Sušice"|"KP133"|"31"
-"Železná Ruda - město, Klatovská na I/27 od Klatov"|"Železná Ruda"|"Železná Ruda"|"KP134"|"2"
-"Losiná I/20, směr Nepomuk"|"Losiná"|"Losiná"|"KP192"|"47"
-"Losiná I/20, směr Plzeň"|"Losiná"|"Losiná"|"KP193"|"47"
-"Starý Plzenec, II/180, Bezručova ulice"|"Starý Plzenec"|"Bezručova ulice"|"KP191"|"48"
-"Stod, Plzeňská I/26"|"Stod"|"Plzeňská I/26"|"KP194"|"49"
-"Nepomuk I/20, směr Blatná"|"Nepomuk"|"Nepomuk"|"KP196"|"50"
-"Nezvěstice I/19, směr Spálené Poříčí"|"Nezvěstice"|"Nezvěstice"|"KP198"|"51"
-"Kaznějov I/20 směr od Plzně"|"Kaznějov"|"Kaznějov"|"KP199"|"52"
-"Kaznějov I/20 směr od Plas"|"Kaznějov "|"Kaznějov"|"KP200"|"52"
-"Kralovice"|"Kralovice "|"Kralovice"|"KP201"|"59"
-"Kralovice"|"Kralovice "|"Kralovice"|"KP202"|"59"
-"Kozolupy II/605 od Plzně"|"Kozolupy "|"Kozolupy"|"KP203"|"53"
-"Zbůch silnice I/26 od Stodu"|"Zbůch "|"Zbůch"|"KP205"|"54"
-"Zbůch silnice I/26 od Plzně"|"Zbůch "|"Zbůch"|"KP206"|"54"
-"Dobřany"|"Dobřany"|"Dobřany"|"KP229"|"55"
-"Plasy"|"Plasy"|"Plasy"|"KP195"|"56"
-"Vejprnice"|"Vejprnice"|"Vejprnice"|"KP197"|"57"
-"Přeštice"|"Přeštice "|"Přeštice"|"KP204"|"58"
