--- app_old/index.php
@@ -1,5 +1,0 @@
-<?php
-
-    header("Location: app/control/control.php");
-
-?>
