--- app_old/readme.txt
@@ -1,1 +1,0 @@
-Tady je pùvodní aplikace. Model se nakopíruje do složky s lumenem.
